Rewrote some nodes to their correct format.
Added a Unity file imported. The importer can import basic nif files with skinned meshes still in development. If you want vertex colors, add a shader that supports them to the mesh and they should show up if they are present.

+++ NiDotNet/NIF/Nodes/QuatKey.cs
@@ -16,7 +16,7 @@ namespace NiDotNet.NIF.Nodes
         {
             if (type != (KeyType) 4)
             {
-                Time = reader.ReadInt32();
+                Time = reader.ReadSingle();
                 Value = (T) Activator.CreateInstance(typeof(T), reader, niFile);
             }
